Assignment 2
(1) Write statements to make legitimate moves starting from the initial configuraion. Remember that,  since `RED_HOME` is 0,  `RED` pieces can only move towards 0, i.e., to points with smaller indices,  whereas `BLACK` pieces move in the opposite direction, from points with smaller to larger indices.  A legitimate move is from a point where the player has checkers to a point that is empty or contains the  same player's checker(s). In addition, if the opponent has _only one_ checker at the target point, it can be  hit, which means that the player's checker replaces it, and the opponent's one moves to the bar.

It is also possible to bear off, i.e., remove one's checkers from the board, but that is only allowed if all the player's checkers have made it to their home board.

If a player has one or more checkers on the bar, then the only legitimate move is to re-enter the game through the opponent's home board.

(2) Write an elementary "interface" to the board that is stored in memory: The user has to enter three numbers (one at a time), first the colour (0 for RED and 1 for BLACK), then the starting point (from 0 to 23), then the endpoint (0 to 23) and let your program make the move that the user entered.  (If the move is not legitimate, there will be an error, but do not bother about that: we do not have the means to check the legitimacy of a move yet.)